backend/requirements.txt
# Core Framework
fastapi==0.110.3
uvicorn[standard]==0.27.0
pydantic==2.6.4
pydantic-settings==2.2.1

# Vector & Graph RAG
pinecone-client==3.0.0
llama-index==0.9.30
neo4j==5.16.0

# Advanced RAG
sentence-transformers==2.2.2
rank-bm25==0.2.2

# Multimodal Processing
unstructured[all]==0.10.30
pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
transformers==4.36.0
torch==2.1.0
pdfplumber==0.10.3
camelot-py==0.11.0
opencv-python==4.8.1.78
python-docx==1.1.2
python-pptx==1.0.2
openpyxl==3.1.5

# LLM & Embeddings
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.6

# Existing (Keep)
boto3==1.28.0
python-dotenv==1.0.1

# Additional utilities
httpx==0.26.0
python-multipart==0.0.6
aiofiles==23.2.1

# Authentication
python-jose[cryptography]==3.3.0
bcrypt==4.1.2

# Rate Limiting
slowapi==0.1.9

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-mock>=3.11.1
